moved article functions into their own controller and model files
