+++ controllers/topic-controller.js
@@ -1,4 +1,4 @@
-const { selectTopics, updateArticle } = require("../models/topic-model");
+const { selectTopics } = require("../models/topic-model");
 
 exports.getTopics = (req, res, next) => {
   selectTopics()
@@ -6,19 +6,6 @@ exports.getTopics = (req, res, next) => {
       res.status(200).send({ topics });
     })
     .catch((err) => {
-      next(err);
-    });
-};
-
-exports.patchArticle = (req, res, next) => {
-  const { article_id } = req.params;
-  const input = req.body;
-  updateArticle(article_id, input)
-    .then((article) => {
-      res.status(201).send({ article });
-    })
-    .catch((err) => {
-      console.log(err);
       next(err);
     });
 };
